Implementing simpler semantics for brush-selection.
Beginning a brushing action without the shift key now immediately clears the
currently selection upon mousedown.

Beginning it with the shift key leaves the selection intact.

Dragging the pointer causes no changes to rendering.

Releasing the button will add the nodes covered by the selector box to the
current selection.

This is to say, shift-brushing can never remove nodes from the selection, and
brushing will always create a new selection from scratch.

This follows the semantics of, for example, selecting drawings in a Google
Slide.