Prevent goofy scrolling effects in certain browsers.
e.g., Chrome on OSX does an "elastic bump" type of animation when attempting to
scroll past the top or bottom of a document.  The effect is jarring when the
intent is for scrolling to accomplish a zoom, so we independently disable scroll
on the document itself, while reinstating it via callback on the SVG element.