Ungrouping works more or less
When ungrouping, the nodes constituting the group node are brought back into the
graph by doing an "add neighborhood with radius 1" operation.  This means, in
general, that the state of the graph before and after a group-ungroup sequence
will not be equal.  We need an "add node" method that doesn't bring in a whole
neighborhood, but just allows the current graph to "absorb" the new node.