Add a gulp task to build documentation.
This commit also adds a specialized .gitignore file to tangelo/docs. The purpose
of the file is to cause the (almost) empty directory to appear, so that when
Tangelo is started without documentation having first been built, the docs
plugin is able to load.

At any time, building the documentation will cause it to show up in the
application's "docs" plugin.