Increase color and size options.
Previously, some colors were read from options but not used.  Now, these options are honored.

If an object is passed to options, and the owner alters certain options used for cola values (such as linkDistance), then calls view.cola.start(), the changed options will take effect.  Colors require notifying the d3 part of things to update.

The transform would get reset any time render() was called, even when it was just an update.  There is now a updateTransform function which will reapply the transform (if, for instance, the transform was supplied by the options and was modified outside of Clique).