Speed up deselection of nodes and links.
Whenever a mouse down event occured, all links and all nodes, regardless of state, were explicitly deselected.  For each node and link, and event was triggered that it had been removed from the selection.  For even modest sized graphs of a few hundred nodes, this results in a very slow process.  Clicking in a black space on the graph multiple times would bind up the web page in copious callbacks (which would also trigger many GC cycles).

This has been changed to only deselect selected nodes and links, and only to rerender nodes if a node was deselected.  remove events are only fired for nodes and links that were actually deselected.